Moved prototype definitions to test folder
